add IsContactExist Feature

+++ Find Contact/Program.cs
@@ -72,14 +72,22 @@ namespace Contactspresentation_layer
 
         static void testDeleteContact(int ID)
         {
-            if (ClsContact.DeleteContact(ID)) 
+            if (ClsContact.IsContactExist(ID))
             {
-                Console.WriteLine("Contact Deleted Successfully");
+                if (ClsContact.DeleteContact(ID))
+                {
+                    Console.WriteLine("Contact Deleted Successfully");
+                }
+                else
+                {
+                    Console.WriteLine("Failed To Delete Contact");
+                }               
             }
             else
             {
-                Console.WriteLine("Failed To Delete Contact");
+                Console.WriteLine("No, this Contact is Not Exist For Delete");
             }
+          
         }
 
         static void testGetAllContacts()
@@ -92,13 +100,27 @@ namespace Contactspresentation_layer
             }
         }
 
+        static void testIsContactExist(int ID)
+        {
+            if (ClsContact.IsContactExist(ID)) 
+            {
+                Console.WriteLine("Yes, Contact Is Found");
+            }
+            else
+            {
+                Console.WriteLine("No, this Contact is Not Exist");
+            }
+        }
+
         static void Main(string[] args)
         {
             //testFindContact(3);
             //testAddNewContact();
             //testUpdateContact(2);
-            //testDeleteContact(20);
-            testGetAllContacts();
+            testDeleteContact(201);
+            //testGetAllContacts();
+            //testIsContactExist(133);
+
             Console.ReadKey();
         }
     }
